07-streaming/study/pyproject.toml
[project]
name = "study"
version = "0.1.0"
description = "PyFlink Stream Processing Study"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kafka-python>=2.3.0",
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=23.0.1",
]
